[nrf fromlist] platform: Add platform default UART control definition
Add platform default UART control definition.
For platforms that wishes to use a UART configuration different from
the default CMSIS control value the default valuse need to be supplied
in the call to set control.

This makes it possible for a platform to set for example Hardware
Flow Contol configuration.
Having support for Hardware Flow Control is useful for running the TF-M
regression tests reliably.

Upstream PR:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/20374

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>