[nrf noup] Include PSA config file in TF-M PSA interface header
Include the PSA config file in TF_M PSA interface header.

This is a noup commit because it is and adapted change based on upstream
commit that contains other changes that would cause merge conflicts.
An additional PR was made to upstream to improve the way the config
header is included.

Upstream commit:
34a0ffd0d14bf98c8dc517e3982c4ed8cec9f56f
Upstream PR:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/21759

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>